Refactor: 코드 구조 재구성
